Rename: components/api 에서 components/input으로
