+++ pages/community/index.vue
@@ -8,17 +8,17 @@
     </v-row>
   </div>
 </template>
- 
+
 <script>
-import TextEditor from '@/components/api/quil/TextEditor.vue'
+import TextEditor from '@/components/input/quil/TextEditor.vue'
 import headMixin from '@/mixins/common/head.js'
 export default {
   components: {
-    TextEditor,
+    TextEditor
   },
   mixins: [headMixin],
-  head () {
+  head() {
     return this.getHead({ pageName: 'community' })
-  }  
+  }
 }
 </script>
